Updates purescript-vim and overrides some purescript indentation levels
The new plugin version fixes the indentation of function definitions after a type signature :D

I also overrode some of the plugin's default indentation settings to match the following
formatting conventions:

    if bool
      then ...
      else ...

    case xs of
      [] -> ...
      (y:ys) -> ...

    bar
      where
        bar = ...

    do
      x <- ...

Note that in the final case this really only applies to `do` on its own line. When writing `do` at
the end of a previously line it will still indent 2 from the *start* of that line, not from the `do`.