Adds Tutor section for Vimstone
This begins the Tutor section, which can be accessed by running `:Tutor
vimstone`. Most of the README moved into the tutor section as it seems like
it would be more useful there. The intention is to grow this tutor section
over time.

Techinically some of this may be better suited to `:help`, but for simplicitly
I've left it all in one place. If it grows sufficiently large it may be useful
to distinguish between fully-fledged help vs. tutorials.