Ignores init.vim settings when loading from VSCode
Some plugins and settings work with VSCode's native nvim integration
while others do not. This simply avoids loading anything in `init.vim`
when nvim is being used as a backend for VSCode editing. We may be able
to construct an alternative config that works well with VSCode if we
decide it's worth it.