Adds some coc.vim mappings
This adds a few more mappings from the "default" coc.vim configuration
in their readme. I have verified this particular mappings work, at least
when compiling with `-haddock`, as is suggested by the
`haskell-language-server` docs. I tried a few other mappings and they
appeared to not be supported by hls, so I did not include them.

I also removed `LanguageClient-neovim`. I thought `coc.vim` was
dependent on it, but that was incorrect, so we don't need it.