Adds which-key plugin to make mappings discoverable.
This adds space-vims which-key plugin and configures names for the
current mappings we have to make the menu look nice.

I tweaked a few of the mappings along the way to try organize things to
be discoverable, but tried to keep things as consistent as possible.

which-key did not play nicely with our <C-W> prefix mapping that led to
all the window commands, so I mapped a bunch explicitly instead which
also let me set names for them in the menu.