Fixes <Space>sa mapping that breaks with which-key defined mappings
I thought I had tested this, must I must not have tested it correctly.

Because this mapping is to an incomplete command (i.e. we explicitly do not
want a carriage return on the end) it cannot be configured using which-key's
configuration syntax.

This us `nnoremap` for the `sa` mapping only, leaving the other which-key
based mappings in place.