Adds BufferFileBrowse() fn and key mapping
This adds a shortcut that launches Telescope's `file_browser` picker
scoped to the directory of the currently active buffer. This makes for
faster searching if you're just looking for things that are your current
modules descendents *and* allows for creation of sibling files easily.

To create a file with `file_browser`

- Launch if with SPC-f-b
- Type the file name in
- Press CTRL-e

Otherwise files can be searched for similar to the normal fuzzy finder
(bound to SPC-a) except child directories won't be searched.