Updates plugins to latest versions
This assumes Neovim 0.7.2 is being used to correctly correspond to the
lsp plugin version.

Most of the plugins were updated to the latest commit on the main
development branch. A few were updated to the the lastest tagged release
or the latest commit with a passing build.

NERDTree is no longer maintained. It has been replaced with `vim-tree`.

The completion plugin has been deprecated by its author and replaced
with a new, lua based version.

This also includes a change to no longer _automatically_ show the error
message inline for diagnostics. You can access the diagnostic via
`<SPACE>el` ("error, line"). This resolves the window display conflict
between errors and documentation.