Completion mapping confirmation set to false
Completion mapping confirmation is now set to false, so that only
explicitly selected items are now confirmed.

This change was necessary due to the fact that the completion menu would
sometimes appear at a time when you wanted to hit enter to go to a new
line. When set to true, the item at the top of the list was selected,
even though it wasn't explicitly selected, and this would prevent you
from entering a new line.