Allows for per-project & global HLS settings
`nvm-lspconfig` allows for a callback to be provided to customize the
language server configuration based on the root directory that it is
being launch for.

This uses that callback for the HLS configuration, allowing for a
system-installed HLS to be used when there is no local HLS script in the
project. It also only settings the formatting provider if a
`fourmolu.yaml` file is present in the project. We can add other
formatting providers as desired based on the presence of their config
files as well.